Add dependencies

+++ OpERP/src/main/java/devopsdistilled/operp/server/data/entity/Item.java
@@ -10,31 +10,31 @@ import javax.persistence.Id;
 @Entity
 public class Item implements Serializable {
 
-    private static final long serialVersionUID = 7751126479626962944L;
-    private  int id;
-    private String name;
+	private static final long serialVersionUID = 7751126479626962944L;
+	private Long id;
+	private String name;
 
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    public int getId() {
-	return id;
-    }
+	@Id
+	@GeneratedValue(strategy = GenerationType.AUTO)
+	public Long getId() {
+		return id;
+	}
 
-    public void setId(int id) {
-	this.id = id;
-    }
+	public void setId(Long id) {
+		this.id = id;
+	}
 
-    public String getName() {
-	return name;
-    }
+	public String getName() {
+		return name;
+	}
 
-    public void setName(String name) {
-	this.name = name;
-    }
+	public void setName(String name) {
+		this.name = name;
+	}
 
-    @Override
-    public String toString() {
-	return new String("Id: " + getId() + "\nName: " + getName());
-    }
+	@Override
+	public String toString() {
+		return new String("Id: " + getId() + "\nName: " + getName());
+	}
 
 }
